+++ ember-data-types/q/record-data-json-api.ts
@@ -1,4 +1,4 @@
-import type { CollectionResourceRelationship, SingleResourceRelationship } from './ember-data-json-api';
+import type { CollectionResourceRelationship, Links, Meta, SingleResourceRelationship } from './ember-data-json-api';
 import type { Dict } from './utils';
 
 /**
@@ -10,12 +10,11 @@ export type AttributesHash = Dict<unknown>;
 export interface JsonApiResource {
   id?: string | null;
   type?: string;
-  attributes?: AttributesHash;
-  relationships?: {
-    [key: string]: SingleResourceRelationship | CollectionResourceRelationship;
-  };
-  meta?: any;
   lid?: string;
+  attributes?: AttributesHash;
+  relationships?: Dict<SingleResourceRelationship | CollectionResourceRelationship>;
+  meta?: Meta;
+  links?: Links;
 }
 
 export interface JsonApiValidationError {
