+++ packages/-ember-data/tests/integration/adapter/store-adapter-test.js
@@ -21,8 +21,7 @@ function moveRecordOutOfInFlight(record) {
   let { store, _internalModel } = record;
   // TODO this would be made nicer by a cancellation API
   let pending = store.getRequestStateService().getPendingRequestsForRecord(_internalModel.identifier);
-  pending.splice(0, pending.length); // release the requests
-  _internalModel.adapterDidError(new Error());
+  pending.splice(0, pending.length);
 }
 
 module('integration/adapter/store-adapter - DS.Store and DS.Adapter integration test', function (hooks) {
