+++ packages/-ember-data/tests/integration/record-data/record-data-errors-test.ts
@@ -27,6 +27,9 @@ class TestRecordIdentifier implements NewRecordIdentifier {
 }
 
 class TestRecordData implements RecordData {
+  getErrors(recordIdentifier: RecordIdentifier): JsonApiValidationError[] {
+    throw new Error('Method not implemented.');
+  }
   id: string | null = '1';
   clientId: string | null = 'test-record-data-1';
   modelName = 'tst';
@@ -299,69 +302,5 @@ module('integration/record-data - Custom RecordData Errors', function (hooks) {
     assert.strictEqual(person.get('errors').errorsFor('name').length, 0, 'no errors on name');
     let lastNameError = person.get('errors').errorsFor('lastName').get('firstObject');
     assert.strictEqual(lastNameError.attribute, 'lastName', 'error shows up on lastName');
-  });
-
-  test('Record data which does not implement getErrors still works correctly with the default DS.Model', async function (assert) {
-    assert.expect(4);
-
-    const personHash = {
-      type: 'person',
-      id: '1',
-      attributes: {
-        name: 'Scumbag Dale',
-      },
-    };
-    let { owner } = this;
-
-    class LifecycleRecordData extends TestRecordData {
-      commitWasRejected(recordIdentifier, errors) {
-        assert.strictEqual(errors[0].detail, 'is a generally unsavoury character', 'received the error');
-        assert.strictEqual(errors[0].source.pointer, '/data/attributes/name', 'pointer is correct');
-      }
-    }
-
-    let TestStore = Store.extend({
-      createRecordDataFor(modelName, id, clientId, storeWrapper) {
-        return new LifecycleRecordData();
-      },
-    });
-
-    let TestAdapter = EmberObject.extend({
-      updateRecord() {
-        return Promise.reject(
-          new InvalidError([
-            {
-              title: 'Invalid Attribute',
-              detail: 'is a generally unsavoury character',
-              source: {
-                pointer: '/data/attributes/name',
-              },
-            },
-          ])
-        );
-      },
-
-      createRecord() {
-        return Promise.resolve();
-      },
-    });
-
-    owner.register('service:store', TestStore);
-    owner.register('adapter:application', TestAdapter, { singleton: false });
-
-    store = owner.lookup('service:store');
-
-    store.push({
-      data: [personHash],
-    });
-    let person = store.peekRecord('person', '1');
-    await person.save().then(
-      () => {},
-      (err) => {}
-    );
-
-    assert.false(person.get('isValid'), 'rejecting the save invalidates the person');
-    let nameError = person.get('errors').errorsFor('name').get('firstObject');
-    assert.strictEqual(nameError.attribute, 'name', 'error shows up on name');
   });
 });
